[ocebuild-cli] Move CLI to new project namespace
