+++ ocebuild_cli/__init__.py
@@ -5,3 +5,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
 ##
 """Methods and classes used for the OCE-Build CLI"""
+
+# Share the same version as the main package.
+from ocebuild.version import *
